Added to json template
Added 5 more people to the json template, for a total of 6. I also added a "timePosted" section because I know we wanted to say what time people posted these. Additionally, I will upload more photos that correspond with the posts. Also note that some image photos are "null" because they don't have to have a photo and I want to make sure we design with this in mind.